feat: refactor extract restaurant_tour to a separated file and rename the restaurant_view_model
